Refactor views module into views package
