+++ webapp/apps/comp/views/__init__.py
@@ -1,0 +1,3 @@
+from .views import RouterView, EditInputsView, OutputsDownloadView, OutputsView
+
+from .api import InputsAPIView, CreateAPIView, DetailAPIView, OutputsAPIView
